Add dead XP mechanic (read description)
XP will no longer disappear. Instead, if it isn't picked up in time it will lose 50% of it's value (signified by a color change from blue to purple) and automatically magnetize towards the player. This way there's no way to lose out on XP, but you won't receive it's full value if you leave it be.